Revert commit ddde9e383ac74f8f4a97603932d11918ee263545

+++ examples/threads/threads.rs
@@ -2,7 +2,7 @@
 
 use std::thread;
 
-static NTHREADS: usize = 10;
+static NTHREADS: i32 = 10;
 
 // This is the `main` thread
 fn main() {
